Move dispatch of generation/update to ConfigFile
With respect to .flux.yaml files, two types control the generation and
update of manifests in a repo:

 - ConfigFile, which is a directory subject to .flux.yaml; and
 - ConfigAware, which represents a set of directories, some of which
   have a ConfigFile, and some of which just have YAML files

The dispatch of operations based on whether a config file has a
PatchUpdated or CommandUpdated field more properly belongs among the
ConfigFile methods, but hitherto has been somewhat scattered.

This commit moves all the detail of how things are generated or
updated into methods of ConfigFile, and just delegates to those from
ConfigAware when there is a config file in effect. The methods to be
delegated to are exported; helpers are not exported.